update db instead of tearing down and rebuilding
